Add Callback handling and tests
A new Callback handling mechanism was added to the SwedbankPay SDK, including the integration of this mechanism in the CheckOutController as well as relevant tests. The CallbackInfo and PaymentOrderInfo classes were introduced to provide an appropriate structure for callback data deserialization. A test was also added to ensure that callbacks get deserialized correctly.