Added ability to create recurring payments (#3)
* Added ability to create recurring payments

* Add functionality for unscheduled payments

This commit introduces the option for creating unscheduled payments in the SwedbankPay SDK. Additional actions such as support creation of unscheduled payments were made in the sample AspNetCore project. Specific buttons and actions for recurring and unscheduled payments have been implemented in the order handling section of the sample project.

* Add PaymentOrderLink to CheckoutController

The PaymentOrderLink has been added to the CheckOutController. This link is assigned with the current payment order's Id, allowing for tracking and usage while in the checkout process.

* Add generatePaymentToken option to checkout

This commit introduces a new option when creating a payment order, called "generatePaymentToken". Users will now see this option in the checkout area, providing the ability to generate a payment token for their transactions. The associated logic has been implemented in the CheckOut Controller and reflected in the relevant methods.