Add token handling to AspNetCore sample application (#4)
* Add token handling to AspNetCore sample application

Introduced a new TokensController and TokenViewModel to manage tokens for the SwedbankPaySdk sample AspNetCore application. Also integrated an IsDeleted field for UserTokenDto and UserTokenResponseItemDto to track token state. A navigation link for "Tokens" has been added to _Layout.cshtml.

* Add functionality to delete unscheduled tokens

New code was added to allow users to delete unscheduled tokens in the payment interface. Changes were implemented across several files, including view and controller files, to facilitate the complete workflow. An action to handle token removal requests was also added into the payment operations.

* Remove unused exception variable in TokensController

The exception variable 'e' in the catch clause of the TokensController was unused. To clean up the code and improve readability, this unused variable has been removed.

* Mapped IsDeleted on token

* Changed IsDeleted to nullable

* Nullabled on Isdeleted

---------

Co-authored-by: Daniel Sundqvist <daniel.sundqvist@authority.se>