Refactor operations base and clean unused code
In this commit, the HttpOperation indexer in the OperationsBase class was refactored to enhance readability. Next, in the PaymentOrdersResource.cs and PaymentController.cs classes, some unused lines of code were removed to reduce clutter. Such improvements aim to maintain the code tidy, improve readability in the codebase, and consequently ease future maintenance.