Update payment order functionality and improve UI
Refactored the payment order functionality in SDK and updated the UI in Sample.AspNetCore. Changes include making the PayeeId in PayeeInfo optional, adding new identifiers to the payment instruments, and reorganizing the payment details display in the Tokens page. Also, added the functionality in the SDK for generating payment tokens and corresponding changes in the API.