Refactor payment token generation in CheckOutController
The code has been refactored to improve the process of generating the payment token in the CheckOutController. An additional conditional statement has been introduced to enable a more secure generation of the payment token. The changes also ensure that the 'DisableStoredPaymentDetails' and 'EnablePaymentDetailsConsentCheckbox' properties are set correctly whenever a payment token is generated.