Add Subsite and SiteId to PayeeInfo (#635)
* Add Subsite and SiteId to PayeeInfo

The PayeeInfo class and related interfaces and DTOs have been extended with Subsite and SiteId properties. The IPayeeInfo interface has been removed and its functionality moved to the PayeeInfo record. The changes provide a more consistent and useful model for working with payee information.

* Refactor PayeeInfo class and add SiteId property

This update refactors the PayeeInfo class, removing unnecessary spaces, and improving the explanation of the Subsite property. A new property, SiteId has been added for handling split settlement transactions. The SiteId is used to specify the sub-merchant for AMEX transactions.