Resolves #654 Add ProductCategory property to PayeeInfo (#656)
* Add ProductCategory property to PayeeInfo

Introduced the ProductCategory property to the PayeeInfo model, DTOs, and its usage in the CheckOutController. This field allows merchants to specify a product category or number for settlement purposes.

* Update action version in test deploy workflow

Upgraded the potiuk/get-workflow-origin action from v1_1 to v1_6. This ensures compatibility with the latest features and improvements in the action.